+++ backend/routes/api.php
@@ -6,7 +6,9 @@ use App\Http\Controllers\Auth\AuthController;
 use App\Http\Controllers\SeatMapController;
 use App\Http\Controllers\VenueController;
 use App\Http\Controllers\Admin\AdminReservationController;
+use App\Http\Controllers\Admin\AdminAccountController;
 use App\Http\Controllers\Client\ClientReservationController;
+use App\Http\Middleware\AdminAccess;
 
 /*
 |--------------------------------------------------------------------------
@@ -31,13 +33,20 @@ Route::prefix('admin')->group(function () {
     Route::post('/login', [AuthController::class, 'adminLogin']);
 });
 
+Route::prefix('admin/accounts')->group(function () {
+    Route::get('/', [AdminAccountController::class, 'index'])->middleware(AdminAccess::class . ':manage_accounts');
+    Route::post('/', [AdminAccountController::class, 'store'])->middleware(AdminAccess::class . ':manage_accounts');
+    Route::put('/me', [AdminAccountController::class, 'updateProfile'])->middleware(AdminAccess::class . ':view_admin');
+    Route::put('/{admin}', [AdminAccountController::class, 'update'])->middleware(AdminAccess::class . ':manage_accounts');
+});
+
 // Venue routes
 Route::prefix('venues')->group(function () {
     Route::get('/', [VenueController::class, 'index']);
     Route::get('/{id}', [VenueController::class, 'show']);
-    Route::post('/', [VenueController::class, 'store']);
-    Route::put('/{id}', [VenueController::class, 'update']);
-    Route::delete('/{id}', [VenueController::class, 'destroy']);
+    Route::post('/', [VenueController::class, 'store'])->middleware(AdminAccess::class . ':manage_venues');
+    Route::put('/{id}', [VenueController::class, 'update'])->middleware(AdminAccess::class . ':manage_venues');
+    Route::delete('/{id}', [VenueController::class, 'destroy'])->middleware(AdminAccess::class . ':manage_venues');
     
     // Additional venue endpoints
     Route::get('/wing/{wing}', [VenueController::class, 'getByWing']);
@@ -60,15 +69,15 @@ Route::get('/rooms/{venueId}/seats', [SeatMapController::class, 'getSeatmapById'
 
 // Admin reservation routes
 Route::prefix('admin/reservations')->group(function () {
-    Route::get('/', [AdminReservationController::class, 'index']);
-    Route::get('/stats', [AdminReservationController::class, 'getStats']);
-    Route::post('/', [AdminReservationController::class, 'store']);
-    Route::get('/{id}', [AdminReservationController::class, 'show']);
-    Route::put('/{id}', [AdminReservationController::class, 'update']);
-    Route::patch('/{id}/approve', [AdminReservationController::class, 'approve']);
-    Route::patch('/{id}/reject', [AdminReservationController::class, 'reject']);
-    Route::patch('/{id}/revert', [AdminReservationController::class, 'revert']);
-    Route::delete('/{id}', [AdminReservationController::class, 'destroy']);
+    Route::get('/', [AdminReservationController::class, 'index'])->middleware(AdminAccess::class . ':view_admin');
+    Route::get('/stats', [AdminReservationController::class, 'getStats'])->middleware(AdminAccess::class . ':view_admin');
+    Route::post('/', [AdminReservationController::class, 'store'])->middleware(AdminAccess::class . ':manage_reservations');
+    Route::get('/{id}', [AdminReservationController::class, 'show'])->middleware(AdminAccess::class . ':view_admin');
+    Route::put('/{id}', [AdminReservationController::class, 'update'])->middleware(AdminAccess::class . ':adjust_reservation_details');
+    Route::patch('/{id}/approve', [AdminReservationController::class, 'approve'])->middleware(AdminAccess::class . ':manage_reservations');
+    Route::patch('/{id}/reject', [AdminReservationController::class, 'reject'])->middleware(AdminAccess::class . ':manage_reservations');
+    Route::patch('/{id}/revert', [AdminReservationController::class, 'revert'])->middleware(AdminAccess::class . ':manage_reservations');
+    Route::delete('/{id}', [AdminReservationController::class, 'destroy'])->middleware(AdminAccess::class . ':delete_reservations');
 });
 
 // Client reservation routes
@@ -79,5 +88,5 @@ Route::prefix('reservations')->group(function () {
     Route::put('/{id}', [ClientReservationController::class, 'update']);
     Route::patch('/{id}/reject', [ClientReservationController::class, 'reject']);
     Route::delete('/{id}', [ClientReservationController::class, 'destroy']);
-    Route::post('/{reservation}/notify', [ClientReservationController::class, 'notify']);
+    Route::post('/{reservation}/notify', [ClientReservationController::class, 'notify'])->middleware(AdminAccess::class . ':manage_reservations');
 });
